server/source: streamline media source search
